+++ src/icemac/ab/calexport/masterdata/tests/test_export.py
@@ -1,71 +1,71 @@
-import icemac.ab.calexport.testing
+from icemac.ab.calendar.interfaces import IEvent
+from icemac.ab.calexport.testing import get_masterdata
+from icemac.addressbook.interfaces import IEntity
+from mechanize import LinkNotFoundError, HTTPError
+import pytest
 
 
-class MasterdataTests(icemac.ab.calexport.testing.BrowserTestCase):
-    """Testing ..export.Masterdata."""
-
-    def test_form_saves_data(self):
-        self.create_category(u'foo')
-        self.create_category(u'bar')
-        browser = self.get_browser('cal-export-editor')
-        browser.open('http://localhost/ab/@@calendar-masterdata.html')
-        browser.getLink('Configure export').click()
-        form_url = ('http://localhost/ab/++attribute++calendar/'
-                    '@@edit-export-masterdata.html')
-        self.assertEqual(form_url, browser.url)
-        browser.getControl('HTML to be inserted above').value = '<html>cal'
-        browser.getControl('foo').selected = True
-        browser.getControl('Apply').click()
-        self.assertEqual(['Data successfully updated.'],
-                         browser.get_messages())
-        browser.open(form_url)
-        self.assertEqual('<html>cal',
-                         browser.getControl('HTML to be inserted above').value)
-        self.assertTrue(browser.getControl('foo').selected)
-        self.assertFalse(browser.getControl('bar').selected)
+def test_export__Masterdata__1(address_book, CategoryFactory, browser):
+    """It saves the entered data."""
+    CategoryFactory(address_book, u'foo')
+    CategoryFactory(address_book, u'bar')
+    browser.login('cal-export-editor')
+    browser.open(browser.CALENDAR_MASTERDATA_URL)
+    browser.getLink('Configure export').click()
+    assert browser.CALEXPORT_MASTER_DATA_URL == browser.url
+    browser.getControl('HTML to be inserted above').value = '<html>cal'
+    browser.getControl('foo').selected = True
+    browser.getControl('Apply').click()
+    assert 'Data successfully updated.' == browser.message
+    browser.open(browser.CALEXPORT_MASTER_DATA_URL)
+    assert '<html>cal' == browser.getControl('HTML to be inserted above').value
+    assert browser.getControl('foo').selected
+    assert not browser.getControl('bar').selected
 
 
-class MasterdataSecurityTests(icemac.ab.calexport.testing.BrowserTestCase):
-    """Security testing ..export.Masterdata."""
-
-    def test_exporter_is_not_able_to_access_fields(self):
-        from mechanize import LinkNotFoundError, HTTPError
-        browser = self.get_browser('cal-exporter')
-        browser.open('http://localhost/ab/@@calendar-masterdata.html')
-        with self.assertRaises(LinkNotFoundError):
-            browser.getLink('Configure export')
-        # The URL is not accesible, too:
-        with self.assertRaises(HTTPError) as err:
-            browser.open('http://localhost/ab/++attribute++calendar/'
-                         '@@edit-export-masterdata.html')
-        self.assertEqual('HTTP Error 403: Forbidden', str(err.exception))
+def test_export__Masterdata__2(address_book, browser):
+    """There is no link rendered to it for users with the exporter role."""
+    browser.login('cal-exporter')
+    browser.open(browser.CALENDAR_MASTERDATA_URL)
+    with pytest.raises(LinkNotFoundError):
+        browser.getLink('Configure export')
 
 
-class ExportMasterDataTests(icemac.ab.calexport.testing.ZODBTestCase):
-    """Testing ..export.ExportMasterData."""
+def test_export__Masterdata__3(address_book, browser):
+    """It cannot be accessed for users with the exporter role."""
+    browser.login('cal-exporter')
+    with pytest.raises(HTTPError) as err:
+        browser.open(browser.CALEXPORT_MASTER_DATA_URL)
+    assert 'HTTP Error 403: Forbidden' == str(err.value)
 
-    def makeOne(self):
-        from ..export import ExportMasterData
-        return ExportMasterData()
 
-    def test_special_field_stores_and_returns_a_field_reference(self):
-        md = self.makeOne()
-        # None by default
-        self.assertIsNone(md.special_field)
-        field = self.create_special_field(md)
-        self.assertEqual(field, md.special_field)
-        # special_field is a property which stores a field reference
-        self.assertEqual(
-            'IcemacAbCalendarEventEvent###Field-1', md._special_field)
-        # Reset to None is possible, too
-        md.special_field = None
-        self.assertIsNone(md.special_field)
-        self.assertIsNone(md._special_field)
+# class ExportMasterDataTests(icemac.ab.calexport.testing.ZODBTestCase):
+#     """Testing ..export.ExportMasterData."""
 
-    def test_special_field_does_not_break_on_a_deleted_field(self):
-        from icemac.addressbook.interfaces import IEntity
-        from icemac.ab.calendar.interfaces import IEvent
-        md = self.makeOne()
-        field = self.create_special_field(md)
-        IEntity(IEvent).removeField(field)
-        self.assertIsNone(md.special_field)
+#     def makeOne(self):
+#
+#         return ExportMasterData()
+
+
+def test_export__ExportMasterData__special_field__1(
+        address_book, SpecialFieldFactory):
+    """It stores and returns a field reference."""
+    md = get_masterdata(address_book)
+    # None by default
+    assert None is md.special_field
+    field = SpecialFieldFactory(address_book)
+    assert field == md.special_field
+    # `special_field` is a property which stores a field reference
+    assert 'IcemacAbCalendarEventEvent###Field-1' == md._special_field
+    # Reset to None is possible, too
+    md.special_field = None
+    assert None is md.special_field
+    assert None is md._special_field
+
+
+def test_export__ExportMasterData__special_field__2(
+        address_book, SpecialFieldFactory):
+    """It does not break on a deleted field."""
+    field = SpecialFieldFactory(address_book)
+    IEntity(IEvent).removeField(field)
+    assert None is get_masterdata(address_book).special_field