+++ src/icemac/ab/calexport/tests/test_roles.py
@@ -1,13 +1,12 @@
-import icemac.ab.calexport.testing
+from icemac.addressbook.principals.roles import has_editor_role
+from icemac.addressbook.principals.roles import has_visitor_role
 
 
-class RoleTests(icemac.ab.calexport.testing.ZCMLTestCase):
-    """Testing ..roles."""
+def test_roles__1(zcmlS):
+    """The calendar export editor role is registered as an editor role."""
+    assert has_editor_role(['icemac.ab.calexport.ExportEditor'])
 
-    def test_calexport_editor_is_in_editor_roles(self):
-        from icemac.addressbook.principals.roles import has_editor_role
-        self.assertTrue(has_editor_role(['icemac.ab.calexport.ExportEditor']))
 
-    def test_calendar_exporter_is_in_visitor_roles(self):
-        from icemac.addressbook.principals.roles import has_visitor_role
-        self.assertTrue(has_visitor_role(['icemac.ab.calexport.Exporter']))
+def test_roles__2(zcmlS):
+    """The calendar exporter role is registered as an visitor role."""
+    assert has_visitor_role(['icemac.ab.calexport.Exporter'])