+++ CHANGES.rst
@@ -9,6 +9,9 @@
 
 - Add ability to render a special CSS class for certain events.
 
+- Use `py.test` fixtures in the tests + convert tests to us the class-less
+  approach.
+
 
 1.0.0 (2015-05-24)
 ==================